requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
numpy==2.1.2
opencv-python==4.10.0.84
peewee==3.17.7
pillow==11.0.0
pony==0.7.19
psycopg2==2.9.10
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
vk-api==11.9.9
